Fix issue 24
http://code.google.com/p/avr-uip/issues/detail?id=24

The code for the uip send P function was based on the uip_send function.
The uip_send function takes two pointers to sram adddress space.
The uip_send_P take a pointer to sram and one to prog mem.
So, the address number could be the same.

Good catch  iamalx